pyproject.toml
[project]
name = "asila"
version = "0.1.0"
description = "Verified Information Hub for Governments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "sqlalchemy[asyncio]>=2.0.27",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "alembic>=1.13.1",
    "redis[hiredis]>=5.0.1",
    "pgvector>=0.2.5",
    "GeoAlchemy2>=0.14.7",
    "google-generativeai>=0.4.1",
    "passlib[bcrypt]>=1.7.4",
    "arq>=0.27.0",
    "structlog>=25.5.0",
    "slowapi>=0.1.9",
    "tenacity>=9.1.4",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.0.2",
    "pytest-asyncio>=0.23.5",
]
